+++ core/src/test/scala-2/caliban/Scala2SpecificSpec.scala
@@ -1,12 +1,8 @@
 package caliban
 
-import caliban.CalibanError.ExecutionError
 import caliban.GraphQL._
-import caliban.InputValue.ObjectValue
 import caliban.TestUtils._
-import caliban.Value.{ NullValue, StringValue }
 import caliban.introspection.adt.__DeprecatedArgs
-import caliban.schema.ArgBuilder
 import caliban.schema.SchemaSpec.introspect
 import zio.test.Assertion._
 import zio.test._
@@ -43,30 +39,6 @@ object Scala2SpecificSpec extends DefaultRunnableSpec {
             """{"events":[{"organizationId":7,"title":"Frida Kahlo exhibition"}],"painters":[{"name":"Claude Monet","movement":"Impressionism"}]}"""
           )
         )
-      },
-      test("Scala 2 buildMissing") {
-        sealed abstract class Nullable[+T]
-        case class SomeNullable[+T](t: T) extends Nullable[T]
-        case object NullNullable          extends Nullable[Nothing]
-        case object MissingNullable       extends Nullable[Nothing]
-
-        implicit def nullableArgBuilder[A](implicit ev: ArgBuilder[A]): ArgBuilder[Nullable[A]] =
-          new ArgBuilder[Nullable[A]] {
-            def build(input: InputValue): Either[ExecutionError, Nullable[A]] = input match {
-              case NullValue => Right(NullNullable)
-              case _         => ev.build(input).map(SomeNullable(_))
-            }
-
-            override def buildMissing: Either[ExecutionError, Nullable[A]] = Right(MissingNullable)
-          }
-
-        case class Wrapper(a: Nullable[String])
-
-        val deriviedAB = implicitly[ArgBuilder[Wrapper]]
-
-        assert(deriviedAB.build(ObjectValue(Map())))(equalTo(Right(Wrapper(MissingNullable)))) &&
-        assert(deriviedAB.build(ObjectValue(Map("a" -> NullValue))))(equalTo(Right(Wrapper(NullNullable)))) &&
-        assert(deriviedAB.build(ObjectValue(Map("a" -> StringValue("x")))))(equalTo(Right(Wrapper(SomeNullable("x")))))
       }
     )
 }
