+++ core/src/test/scala-3/caliban/Scala3SpecificSpec.scala
@@ -73,30 +73,6 @@ object Scala3SpecificSpec extends DefaultRunnableSpec {
         assertM(interpreter.flatMap(_.execute(query)).map(_.data.toString))(
           equalTo("""{"item":{"a":1}}""")
         )
-      },
-      test("Scala 3 buildMissing") {
-        sealed abstract class Nullable[+T]
-        case class SomeNullable[+T](t: T) extends Nullable[T]
-        case object NullNullable          extends Nullable[Nothing]
-        case object MissingNullable       extends Nullable[Nothing]
-
-        implicit def nullableArgBuilder[A](implicit ev: ArgBuilder[A]): ArgBuilder[Nullable[A]] =
-          new ArgBuilder[Nullable[A]] {
-            def build(input: InputValue): Either[ExecutionError, Nullable[A]] = input match {
-              case NullValue => Right(NullNullable)
-              case _         => ev.build(input).map(SomeNullable(_))
-            }
-
-            override def buildMissing: Either[ExecutionError, Nullable[A]] = Right(MissingNullable)
-          }
-
-        case class Wrapper(a: Nullable[String])
-
-        val deriviedAB = implicitly[ArgBuilder[Wrapper]]
-
-        assert(deriviedAB.build(ObjectValue(Map())))(equalTo(Right(Wrapper(MissingNullable)))) &&
-        assert(deriviedAB.build(ObjectValue(Map("a" -> NullValue))))(equalTo(Right(Wrapper(NullNullable)))) &&
-        assert(deriviedAB.build(ObjectValue(Map("a" -> StringValue("x")))))(equalTo(Right(Wrapper(SomeNullable("x")))))
       }
     )
 }
